+++ CHANGELOG.md
@@ -6,7 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
-## [0.12.0] - Unreleased
+## [0.12.0] - 2026-07-03
 
 ### Breaking Changes (pre-v1.0 API cleanup)
 
@@ -36,7 +36,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ---
 
-## [0.11.2] - Unreleased
+## [0.11.2] - 2026-07-03
 
 ### Security